# Adapt to GPUCompiler 0.18