--- a/src/compiler/codegen.jl
+++ b/src/compiler/codegen.jl
@@ -143,7 +143,7 @@ function compile(@nospecialize(job::CompilerJob))
     end
 end
 function link(@nospecialize(job::CompilerJob), compiled)
-    Runtime.@log_start(:link, (;f=job.config.source.f, tt=job.config.source.tt), nothing)
+    Runtime.@log_start(:link, (;f=job.source.f, tt=job.source.tt), nothing)
     device = job.config.params.device
     global_hooks = job.config.params.global_hooks
     (;obj, entry, globals) = compiled
@@ -152,7 +152,7 @@ function link(@nospecialize(job::CompilerJob), compiled)
     obj = codeunits(obj)
     exe = AMDGPU.create_executable(device, entry, obj; globals=globals)
     mod = ROCModule(exe)
-    fun = ROCFunction(mod, entry)
+    fun = ROCFunction(mod, entry, hash(job.source, UInt64(0)))
 
     # initialize globals from hooks
     for gname in first.(globals)
@@ -164,18 +164,17 @@ function link(@nospecialize(job::CompilerJob), compiled)
         end
         if hook !== nothing
             @debug "Initializing global $gname"
-            Runtime.@log_start(:global_init, (;f=job.config.source.f, tt=job.config.source.tt, gname), nothing)
+            Runtime.@log_start(:global_init, (;f=job.source.f, tt=job.source.tt, gname), nothing)
             gbl = Runtime.get_global(exe, gname)
             hook(gbl, mod, device)
-            Runtime.@log_finish(:global_init, (;f=job.config.source.f, tt=job.config.source.tt, gname), nothing)
+            Runtime.@log_finish(:global_init, (;f=job.source.f, tt=job.source.tt, gname), nothing)
         else
             @debug "Uninitialized global $gname"
             continue
         end
     end
 
-    Runtime.@log_finish(:link, (;f=job.config.source.f, tt=job.config.source.tt), nothing)
-
+    Runtime.@log_finish(:link, (;f=job.source.f, tt=job.source.tt), nothing)
     return fun
 end