--- a/src/runtime/kernel.jl
+++ b/src/runtime/kernel.jl
@@ -53,6 +53,7 @@ end
 mutable struct ROCFunction
     mod::ROCModule
     entry::String
+    hash::UInt64
 end
 
 ## Kernel instance