- Add caches for compiler config & kernel instances.
- Adapt to GPUCompiler 0.18.
- Use `FunctionSpec` hash as `ROCFunction` hash. It contains `world` so should be fine as a replacement for specialization id, I guess.
- Minor refactor.

Before:
```julia
julia> @benchmark rocfunction(identity, Tuple{Nothing})
BenchmarkTools.Trial: 10000 samples with 194 evaluations.
 Range (min … max):  497.629 ns …  12.318 μs  ┊ GC (min … max): 0.00% … 94.66%
 Time  (median):     514.644 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   531.847 ns ± 373.630 ns  ┊ GC (mean ± σ):  2.69% ±  3.65%

       ▂▁▂▄▆█▆▂                                                  
  ▂▂▃▅█████████▇▅▄▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▃▃▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂ ▃
  498 ns           Histogram: frequency by time          603 ns <

 Memory estimate: 352 bytes, allocs estimate: 9.
```

This PR:
```julia
julia> @benchmark rocfunction(identity, Tuple{Nothing})
BenchmarkTools.Trial: 10000 samples with 197 evaluations.
 Range (min … max):  455.995 ns …  14.574 μs  ┊ GC (min … max): 0.00% … 95.98%
 Time  (median):     464.015 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   474.701 ns ± 326.955 ns  ┊ GC (mean ± σ):  1.67% ±  2.35%

     ▂▅▆▇████▇▆▆▅▄▃▃▂▂▂▂▂▁  ▁                 ▁▁ ▁  ▁           ▃
  ▂▆██████████████████████████▇▆▆▇▆▇▆▅▆▆█▇▇███████████▇▇▇▆▆▅▄▅▄ █
  456 ns        Histogram: log(frequency) by time        505 ns <

 Memory estimate: 144 bytes, allocs estimate: 4.
```